--- a/README.md
+++ b/README.md
@@ -11,9 +11,43 @@ npm install
 
 ## Development Server
 
+We are using webpack's DllPlugin, so **we need to build our DLL vendor package before running our standard webpack watch**. To do so, run this once:
+
+```bash
+npm run webpack-dll
+```
+
+This puts all the vendor files in their own bundle so we don't need to scan them when rebuilding our files during development. Now to start the dev server, there are two options:
+
+### Web Server + Webpack Watch in one command
+
 ```bash
 npm run dev
 ```
+
+### Web Server separate from Webpack Watch
+
+Start the web server:
+
+```bash
+npm run start-dev
+```
+
+Start webpack watch:
+
+```bash
+npm run watch-client
+```
+
+### Caching with HardSourceWebpackPlugin
+
+We are using [HardSourceWebpackPlugin](https://github.com/mzgoddard/hard-source-webpack-plugin). If your webpack build is operating strangely, be sure to run a
+
+```bash
+npm run webpack-clean-cache
+```
+
+This plugin dramatically speeds up build times, but does require you to clean the cache occasionally (when problems arise).
 
 ### Using webpack-dashboard
 
@@ -50,7 +84,7 @@ npm run build
 npm run start
 ```
 
-## Deploying 
+## Deploying
 
 The site is currently configured to deploy to http://client-dot-mlab-oti.appspot.com/. To do so, run: