--- a/package.json
+++ b/package.json
@@ -109,7 +109,7 @@
     "express-session": "^1.12.1",
     "extract-text-webpack-plugin": "^0.9.1",
     "file-loader": "^0.9.0",
-    "hard-source-webpack-plugin": "0.0.34",
+    "hard-source-webpack-plugin": "0.0.39",
     "http-proxy": "^1.12.0",
     "jquery": "^3.1.0",
     "json-loader": "^0.5.4",