--- a/package.json
+++ b/package.json
@@ -24,6 +24,7 @@
     "lint": "eslint -c .eslintrc src",
     "start-dev": "better-npm-run start-dev",
     "watch-client": "better-npm-run watch-client",
+    "webpack-dll": "webpack --colors --progress --config webpack/dll.config.js",
     "webpack-dashboard": "WEBPACK_DASHBOARD=true node webpack/webpack-dev-server.js",
     "dev": "concurrently --kill-others \"npm run start-dev\" \"npm run watch-client\"",
     "test": "karma start",