--- a/src/components/ScatterGroup/ScatterGroup.jsx
+++ b/src/components/ScatterGroup/ScatterGroup.jsx
@@ -1,5 +1,6 @@
 import React, { PureComponent, PropTypes } from 'react';
-import { Row, Col } from 'react-bootstrap';
+import Row from 'react-bootstrap/lib/Row';
+import Col from 'react-bootstrap/lib/Col';
 import { metrics } from '../../constants';
 
 import { ScatterPlot, SelectableDropdown } from '../../components';