--- a/src/components/SelectableDropdown/SelectableDropdown.jsx
+++ b/src/components/SelectableDropdown/SelectableDropdown.jsx
@@ -1,5 +1,6 @@
 import React, { PureComponent, PropTypes } from 'react';
-import { DropdownButton, MenuItem } from 'react-bootstrap';
+import DropdownButton from 'react-bootstrap/lib/DropdownButton';
+import MenuItem from 'react-bootstrap/lib/MenuItem';
 
 
 /**