--- a/src/containers/index.js
+++ b/src/containers/index.js
@@ -2,4 +2,3 @@ export App from './App/App';
 export HomePage from './HomePage/HomePage';
 export LocationPage from './LocationPage/LocationPage';
 export NotFoundPage from './NotFoundPage/NotFoundPage';
-export DevTools from './DevTools/DevTools';