Webpack rebuilds were slow. I did a bunch of research and profiling and figured out a few ways to optimize the rebuild times:

- Profile stats with `webpack --config webpack/dev.config.js --profile --json > stats.json` and http://webpack.github.io/analyse/. This led to prefetching base.scss which really cut down build times (surprisingly)
- I scanned the modules in the profile and saw that we were pulling in packages we weren't using, like all the components of react-bootstrap. This is because webpack 1 doesn't understand ES6 module syntax. So `import { Nav } from 'react-bootstrap'` imports all of react-bootstrap. I believe webpack 2 fixes this, but we aren't using it. So instead we do `import Nav from 'react-bootstrap/lib/Nav'`
- I removed the redux devtools from our dependencies since we just use the browser extension anyway and so there's no reason to build it and its dependencies.
- Instead of excluding `node_modules` from babel-loader, I told it to only include `src`. This probably didn't help much, but it was recommended somewhere.
- The big change is adding in webpack's DllPlugin. I experimented with a bunch of ways to separate out our vendor code and not have it rebuild, and this seems to be the way to do it (**not** CommonsChunkPlugin). Essentially you compile a bundle for your vendor files separately and just reference it in your main webpack build. This introduces a new build step `npm run webpack-dll` everytime a new dependency is added (and those dependencies should also be added to `webpack/dll.config.js`. But the rebuild times are worth it.

I'm now seeing webpack builds taking between 400ms and 2s (typically around 1s I thnk), down from 3-6s (typically 4.5s).

So after merging this PR, to get started, run:

```
npm install
npm run webpack-clean-cache
npm run webpack-dll
```

Then do your dev server as normal. For me, that's in two separate terminal windows:

```
npm run start-dev
npm run watch-client
```

Note I ask you to npm install since I updated hard-source-webpack-plugin. I didn't add any new dependencies.