@pbeshai 2016-09-15
This also pointed out that one of our dependencies is pulling in all of `lodash`. :(