@pbeshai 2016-09-15
This will also have implications for the client/production deploy since I modified Html.jsx to include the scripts differently but I didn't test it