@vlandham 2016-09-15
thanks for all that work @pbeshai !

i thought we were getting rid of lodash? i can work towards removing it if we still are using it somewhere. 

Also, looks like our datepicker now no longer requires jquery - so we could update and remove that? https://github.com/airbnb/react-dates/issues/39