@pbeshai 2016-09-15
We did get rid of lodash, but some of our dependencies are using it. :( and jquery doesn't show up in the build right now, so I'm not sure. but we should still update that date plugin anyway.